+++ app/src/main/java/com/epam/training/fooddelivery/converters/FoodModelToFoodConverter.java
@@ -6,6 +6,8 @@ import com.epam.training.fooddelivery.model.FoodModel;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class FoodModelToFoodConverter implements Converter<FoodModel,Food> {
 
@@ -13,8 +15,8 @@ public class FoodModelToFoodConverter implements Converter<FoodModel,Food> {
     public Food convert(FoodModel source){
         Food food = new Food();
         food.setId(source.getId());
-        food.setCalorie(source.getCalorie());
-        food.setPrice(source.getPrice());
+        food.setCalorie(BigDecimal.valueOf(source.getCalorie()));
+        food.setPrice(BigDecimal.valueOf(source.getPrice()));
         food.setCategory(Category.valueOf(source.getCategory().toString()));
         food.setDescription(source.getDescription());
         food.setName(source.getName());
