+++ app/src/main/java/com/epam/training/fooddelivery/converters/FoodToFoodModelConverter.java
@@ -12,8 +12,8 @@ public class FoodToFoodModelConverter implements Converter<Food,FoodModel> {
     public FoodModel convert(Food source){
         FoodModel foodModel = new FoodModel();
         foodModel.setId(source.getId());
-        foodModel.setCalorie(source.getCalorie());
-        foodModel.setPrice(source.getPrice());
+        foodModel.setCalorie(source.getCalorie().longValue());
+        foodModel.setPrice(source.getPrice().longValue());
         foodModel.setCategory(FoodModel.CategoryEnum.valueOf(source.getCategory().toString()));
         foodModel.setDescription(source.getDescription());
         foodModel.setName(source.getName());
